update server side and clear code from unusing comments
